pyproject.toml
[project]
name = "pix4dmatic-mcp"
version = "0.1.0"
description = "Local MCP server for controlling PIX4Dmatic through Windows GUI automation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.2.0",
  "pillow>=10.0.0",
  "psutil>=5.9.0",
  "pydantic>=2.0.0",
  "pywinauto>=0.6.8",
]

[project.scripts]
pix4dmatic-mcp = "pix4dmatic_mcp.server:main"

[build-system]
requires = ["setuptools>=57", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
